Work in progress, 2025/12/07.

+++ src/rc_flatten.rs
@@ -1,4 +1,4 @@
-pub fn rc_flatten(row: i8, col: i8) -> i8
+pub fn rc_flatten(row: usize, col: usize) -> usize
 {
     // This function acts as a translation between the 2D board position and an
     // index of the board array.
